Removed uint64_t (long long int is still 4 bytes when compiling 32 bits)

+++ include/string.h
@@ -3,7 +3,7 @@
 
 u32 strlen(char *str);
 void reverse(char *str);
-void itoa(u64 number, char *str, u32 base);
+void itoa(u32 number, char *str, u32 base);
 
 static char *exceptions_string[32] = {
 "Divide-by-zero",
